Move MutationOptions to namespace
