+++ src/core/deprecated.ts
@@ -1,3 +1,5 @@
+import type { ApolloCache } from "@apollo/client/cache";
+
 import type { ApolloClient } from "./ApolloClient.js";
 import type { OperationVariables } from "./types.js";
 
@@ -6,6 +8,13 @@ export type ApolloClientOptions = ApolloClient.Options;
 
 /** @deprecated Use `ApolloClient.DevtoolsOptions` instead */
 export type DevtoolsOptions = ApolloClient.DevtoolsOptions;
+
+/** @deprecated Use `ApolloClient.MutateOptions` instead */
+export type MutationOptions<
+  TData = unknown,
+  TVariables extends OperationVariables = OperationVariables,
+  TCache extends ApolloCache = ApolloCache,
+> = ApolloClient.MutateOptions<TData, TVariables, TCache>;
 
 /** @deprecated Use `ApolloClient.QueryOptions` instead */
 export type QueryOptions<
