+++ src/core/watchQueryOptions.ts
@@ -177,51 +177,3 @@ export type SubscriptionOptions<
   /** {@inheritDoc @apollo/client!SubscriptionOptionsDocumentation#extensions:member} */
   extensions?: Record<string, any>;
 } & VariablesOption<NoInfer<TVariables>>;
-
-export type MutationOptions<
-  TData = unknown,
-  TVariables extends OperationVariables = OperationVariables,
-  TCache extends ApolloCache = ApolloCache,
-> = {
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#optimisticResponse:member} */
-  optimisticResponse?:
-    | Unmasked<NoInfer<TData>>
-    | ((
-        vars: TVariables,
-        { IGNORE }: { IGNORE: IgnoreModifier }
-      ) => Unmasked<NoInfer<TData>> | IgnoreModifier);
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#updateQueries:member} */
-  updateQueries?: MutationQueryReducersMap<TData>;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#refetchQueries:member} */
-  refetchQueries?:
-    | ((
-        result: NormalizedExecutionResult<Unmasked<TData>>
-      ) => InternalRefetchQueriesInclude)
-    | InternalRefetchQueriesInclude;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#awaitRefetchQueries:member} */
-  awaitRefetchQueries?: boolean;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#update:member} */
-  update?: MutationUpdaterFunction<TData, TVariables, TCache>;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#onQueryUpdated:member} */
-  onQueryUpdated?: OnQueryUpdated<any>;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#errorPolicy:member} */
-  errorPolicy?: ErrorPolicy;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#context:member} */
-  context?: DefaultContext;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#fetchPolicy:member} */
-  fetchPolicy?: MutationFetchPolicy;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#keepRootFields:member} */
-  keepRootFields?: boolean;
-
-  /** {@inheritDoc @apollo/client!MutationOptionsDocumentation#mutation:member} */
-  mutation: DocumentNode | TypedDocumentNode<TData, TVariables>;
-} & VariablesOption<NoInfer<TVariables>>;
